feat: factored out UserSettings, docs
