+++ src/pages/user/[id].tsx
@@ -17,6 +17,11 @@ import { services } from '../../services';
 import { TPost, TUser } from '../../types';
 import { FollowUserButton } from '../../components/FollowUserButton';
 
+/**
+ * @description
+ * This page shows detail of any user and the curent user profile with some additional features.
+ */
+
 type TUserPage = {
 	user: TUser;
 	posts: TPost[];
@@ -78,11 +83,9 @@ const UserPage: FC<TUserPage> = ({ user, posts, likes }: InferGetServerSideProps
 							{user.userName}
 						</IconText>
 					</NextLink>
-
 					<IconText icon="location" colorScheme="slate" size="S">
 						{user.city}
 					</IconText>
-
 					<IconText icon="calendar" colorScheme="slate" size="S">
 						<TimeStamp date={user.createdAt} prefix="Mitglied seit" />
 					</IconText>
